refactor(suite): ConfirmOnDevice component
